% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalDescriptors.R
\name{agrilusExitHoleDensity}
\alias{agrilusExitHoleDensity}
\title{Agrilus exit hole density (m^-2)}
\usage{
agrilusExitHoleDensity(h, d, s = 2)
}
\arguments{
\item{h}{number of Agrilus exit holes}

\item{d}{diameter at breast height (mm)}

\item{s}{height to which stem surveyed from the tree base (m)}
}
\description{
Calculate Agrilus biguttatus exit hole density.
}
