% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalDescriptors.R
\name{bleedPrevalence}
\alias{bleedPrevalence}
\title{Estimated bleed prevalence (%)}
\usage{
bleedPrevalence(a, A, b, B, d, s = 3)
}
\arguments{
\item{a}{average active bleed size (mm)}

\item{A}{number of active bleeds}

\item{b}{average black stain size}

\item{B}{number of black stains}

\item{d}{diameter at breast height (mm)}

\item{s}{height to which stem surveyed from the tree base (m)}
}
\description{
Calculate estimated bleed prevalence.
}
