% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalDescriptors.R
\name{crownCondition}
\alias{crownCondition}
\title{Crown condition}
\usage{
crownCondition(m, t)
}
\arguments{
\item{m}{missing crown (%)}

\item{t}{crown transparency (%)}
}
\description{
Calculate crown condition (%).
}
