% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalDescriptors.R
\name{crownProductionEfficiency}
\alias{crownProductionEfficiency}
\title{Crown production efficiency}
\usage{
crownProductionEfficiency(crown_surface_area, crown_volume)
}
\arguments{
\item{crown_surface_area}{crown surface area (m^2)}

\item{crown_volume}{crown volume (m^3)}
}
\description{
Calculate the crown production efficiency
}
