% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalDescriptors.R
\name{crownSurfaceArea}
\alias{crownSurfaceArea}
\title{Crown surface area (m^2)}
\usage{
crownSurfaceArea(r, h, l, c)
}
\arguments{
\item{r}{crown radius (m)}

\item{h}{total height (m)}

\item{l}{lower crown height (m)}

\item{c}{crown condition (\%)}
}
\description{
Calculate the crown surface area
}
\examples{
crownSurfaceArea(3,15,10,50)
}
