% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalDescriptors.R
\name{crownVolume}
\alias{crownVolume}
\title{Estimated crown volume (m^3)}
\usage{
crownVolume(r, h, l, c)
}
\arguments{
\item{r}{crown radius (m)}

\item{h}{total height (m)}

\item{l}{lower crown height (m)}

\item{c}{crown condition (\%)}
}
\description{
Calculate estimated crown volume.
}
\examples{
crownVolume(3,15,10,50)
}
