% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalDescriptors.R
\name{liveCrownRatio}
\alias{liveCrownRatio}
\title{Live crown ratio}
\usage{
liveCrownRatio(h, l)
}
\arguments{
\item{h}{total height (m)}

\item{l}{lower crown height (m)}
}
\description{
Calculate the live crown ratio
}
\examples{
liveCrownRatio(15,10)
}
