% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minMaxScale.R
\name{minMaxScale}
\alias{minMaxScale}
\title{Min-max scaling}
\usage{
minMaxScale(vec)
}
\arguments{
\item{vec}{vector of numbers to scale}
}
\description{
Variable min-max scaling.
}
\examples{
set.seed(1234)

d <- runif(20,1,10)

minMaxScale(d)
}
