% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypingTemplate.R
\name{phenotypingTemplate}
\alias{phenotypingTemplate}
\title{Phenotyping template}
\usage{
phenotypingTemplate(path = ".")
}
\arguments{
\item{path}{directory path for export output}
}
\description{
Export a copy of the oak phenotyping data collection spreadsheet.
}
\examples{
\dontrun{
phenotypingTemplate()
}
}
