% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPhenotypeSheet.R
\name{readPhenotypeSheet}
\alias{readPhenotypeSheet}
\title{Read phenptyping sheet}
\usage{
readPhenotypeSheet(file)
}
\arguments{
\item{file}{file path to excel file to parse}
}
\description{
Parse .xlsx phenotype data collection sheets.
}
\examples{
library(dplyr)

## Retrieve file paths for example data
files <- list.files(system.file('phenotypeDataCollectionSheets',
  package = 'pdi'),full.names = TRUE)

## Prepare data
d <- readPhenotypeSheet(files[1])
}
