% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{map}
\title{Re-exported functions}
\usage{
lhs \%>\% rhs

map(.x, .f, ..., .progress = FALSE)
}
\description{
These functions are re-exported from other packages.
}
\keyword{internal}
