% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf}
\alias{rf}
\title{rf}
\usage{
rf(analysisTable, cls, nreps = 100, seed = 1234)
}
\arguments{
\item{analysisTable}{tibble of phenotype data suitable for random forest analysis as returned by \code{preparePhenotypeData}}

\item{cls}{analysisTable column to use as response vector. NULL for unsupervised analyses.}

\item{nreps}{number of repetitions}

\item{seed}{random number seed}
}
\description{
perform random forest repetitions
}
