% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteCorrection.R
\name{siteCorrection}
\alias{siteCorrection}
\title{siteCorrection}
\usage{
siteCorrection(
  phenoData,
  descriptors = c("Diameter at breast height (mm)", "Lower crown height (m)",
    "Timber height (m)", "Total height (m)", "Crown radius (m)")
)
}
\arguments{
\item{phenoData}{phenoData tibble containing phenotype data}

\item{descriptors}{columns of phenoData on which to perform site correction}
}
\description{
perform a site correction of selected descriptors
}
