% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteCorrectionFactors.R
\name{siteCorrectionFactors}
\alias{siteCorrectionFactors}
\title{siteCorrectionFactors}
\usage{
siteCorrectionFactors(
  phenoData,
  descriptors = c("Diameter at breast height (m)", "Lower crown height (m)",
    "Timber height (m)", "Total height (m)", "Crown radius (m)")
)
}
\arguments{
\item{phenoData}{phenoData tibble containing phenotype data}

\item{descriptors}{columns of phenoData on which calculate site correction factors}
}
\description{
return site correction correction factors of selected descriptors
}
