% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDIs.R
\name{calcDIs}
\alias{calcDIs}
\title{Calculate Decline Indexes}
\usage{
calcDIs(rfModels, invertPDI = TRUE, invertDAI = TRUE)
}
\arguments{
\item{rfModels}{list containing random forest models}

\item{invertPDI}{invert the PDI scale? TRUE/FALSE}

\item{invertDAI}{invert the DAI scale? TRUE/FALSE}
}
\description{
Calculate Phenotypic Decline Index (PDI) and Decline Acuteness Index (DAI)
}
