% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDIs.R
\name{calcDIs}
\alias{calcDIs}
\title{Calculate Decline Indexes}
\usage{
calcDIs(rfModels, PDI = TRUE, DAI = TRUE, invertPDI = TRUE, invertDAI = TRUE)
}
\arguments{
\item{rfModels}{list containing random forest models as returned by \code{rf()}}

\item{PDI}{TRUE/FALSE, calculate PDI?}

\item{DAI}{TRUE/FALSE, calculate DAI?}

\item{invertPDI}{invert the PDI scale? TRUE/FALSE. Ignored if argument PDI is FALSE}

\item{invertDAI}{invert the DAI scale? TRUE/FALSE. Ignored if argument DAI is FALSE}
}
\description{
Calculate Phenotypic Decline Index (PDI) and Decline Acuteness Index (DAI).
}
