% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDIs.R
\name{calcDIs}
\alias{calcDIs}
\title{Calculate Decline Indexes}
\usage{
calcDIs(files)
}
\arguments{
\item{files}{vector of file paths to phenotype data collection sheets (.xlsx)}
}
\description{
Calculate Phenotypic Decline Index (PDI) and Decline Acuteness Index from phenotypic descriptors.
}
