% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{descriptorContributions}
\alias{descriptorContributions}
\title{descriptorContributions}
\usage{
descriptorContributions(rfModels)
}
\arguments{
\item{rfModels}{list containing random forest models as returned by \code{rf()}}
}
\description{
Calculate average descriptor contributions to random forest models.
}
\details{
See \code{see ?randomForest::importance} for details on random forest importance metrics.
}
