% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnalysisTable.R
\name{makeAnalysisTable}
\alias{makeAnalysisTable}
\title{makeAnalysisTable}
\usage{
makeAnalysisTable(phenoData)
}
\arguments{
\item{phenoData}{tibble containing phenotype data}
}
\description{
prepare phenoData table ready for random forest analysis
}
