% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{mds}
\alias{mds}
\title{Multidimensional scaling}
\usage{
mds(rfModels, dimensions = 2)
}
\arguments{
\item{rfModels}{list containing random forest models as returned by \code{rf()}}

\item{dimensions}{number of dimensions to scale to}
}
\description{
perform multidimensional scaling of random forest proximities
}
