% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPhenotypeSheet.R
\name{readPhenotypeSheet}
\alias{readPhenotypeSheet}
\title{readPhenotypeSheet}
\usage{
readPhenotypeSheet(file)
}
\arguments{
\item{file}{file path to excel file to parse}
}
\description{
parse .xls phenotype data collection sheets
}
\examples{
file <- list.files(system.file('phenotypeDataCollectionSheets',package = 'pdi'),full.names = TRUE)
d <- readPhenotypeSheet(file[1])
}
