% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf}
\alias{rf}
\title{Random forest analysis}
\usage{
rf(analysisTable, cls, params = list(), nreps = 100, seed = 1234)
}
\arguments{
\item{analysisTable}{tibble of phenotype data suitable for random forest analysis as returned by \code{preparePhenotypeData}}

\item{cls}{analysisTable column to use as response vector. NULL for unsupervised analyses.}

\item{params}{additional arguments to pass to randomForest::randomForest}

\item{nreps}{number of repetitions}

\item{seed}{random number seed}
}
\description{
Perform random forest repetitions.
}
