% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteAdjustmentFactors.R
\name{siteAdjustmentFactors}
\alias{siteAdjustmentFactors}
\title{Site adjustment factors}
\usage{
siteAdjustmentFactors(
  phenoData,
  descriptors = c("Diameter at breast height (m)", "Lower crown height (m)",
    "Timber height (m)", "Total height (m)", "Crown radius (m)")
)
}
\arguments{
\item{phenoData}{phenoData tibble containing phenotype data}

\item{descriptors}{columns of phenoData on which calculate site correction factors}
}
\description{
Return site adjustment factors of selected phenotypic descriptors.
}
